OneDate/requirement.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==2.2.16
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.0.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
Markdown==3.3.3
Pillow==8.0.1
pkg-resources==0.0.0
progress==1.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.4.0
